+++ src/change-event-buffer.ts
@@ -1,8 +1,6 @@
 /**
  * a buffer-cache which holds the last X changeEvents of the collection
- * TODO this could be optimized to only store the last event of one document
  */
-
 import {
     Subscription
 } from 'rxjs';
@@ -97,6 +95,7 @@ export class ChangeEventBuffer {
         return changeEvents.slice(0);
         // TODO the old implementation was wrong
         // because it did not correctly reassigned the previousData of the changeevents
+        // this should be added to the event-reduce library and not be done in RxDB
         const docEventMap: any = {};
         changeEvents.forEach(changeEvent => {
             docEventMap[changeEvent.documentId] = changeEvent;