+++ src/overwritable.ts
@@ -6,7 +6,7 @@ import {
     pluginMissing
 } from './util';
 
-const funs = {
+export const overwritable = {
     /**
      * if this method is overwritte with one
      * that returns true, we do additional checks
@@ -40,13 +40,9 @@ const funs = {
      * overwritte to map error-codes to text-messages
      */
     tunnelErrorMessage(message: string): string {
-        // TODO better text with link
         return `RxDB Error-Code ${message}.
         - To find out what this means, use the dev-mode-plugin https://pubkey.github.io/rxdb/custom-build.html#dev-mode
         - Or search for this code https://github.com/pubkey/rxdb/search?q=${message}
         `;
     }
 };
-
-// TODO no default exports
-export default funs;