+++ src/plugins/encryption.ts
@@ -16,7 +16,7 @@ import {
     Crypter
 } from '../crypter';
 import type { RxPlugin, RxDatabase } from '../types';
-import { hash } from '../util';
+import { hash, LOCAL_PREFIX } from '../util';
 
 const minPassLength = 8;
 
@@ -57,7 +57,7 @@ export function storePasswordHashIntoDatabase(
         return Promise.resolve(false);
     }
     const pwHash = hash(rxDatabase.password);
-    return rxDatabase.internalStore.get('_local/pwHash')
+    return rxDatabase.internalStore.get(LOCAL_PREFIX + 'pwHash')
         .catch(() => null)
         .then((pwHashDoc: PasswordHashDocument | null) => {
             /**
@@ -66,7 +66,7 @@ export function storePasswordHashIntoDatabase(
              */
             if (!pwHashDoc) {
                 return rxDatabase.internalStore.put({
-                    _id: '_local/pwHash',
+                    _id: LOCAL_PREFIX + 'pwHash',
                     value: pwHash
                 }).catch(() => null)
                     .then(() => true);